Rename MuiCheckBox --> BloomCheckBox
