+++ src/BloomBrowserUI/bookEdit/copyrightAndLicense/CopyrightPanel.tsx
@@ -7,7 +7,7 @@ import { useEffect, useState } from "react";
 import { kMutedTextGray } from "../../bloomMaterialUITheme";
 import { NoteBox } from "../../react_components/BloomDialog/commonDialogComponents";
 import { Div } from "../../react_components/l10nComponents";
-import { BloomCheckbox } from "../../react_components/muiCheckBox";
+import { BloomCheckbox } from "../../react_components/BloomCheckBox";
 import { MuiTextField } from "../../react_components/muiTextField";
 import { PWithLink } from "../../react_components/pWithLink";
 import { IDerivativeInfo } from "./CopyrightAndLicenseDialog";
