+++ src/BloomBrowserUI/publish/ReaderPublish/BulkBloomPub/BulkBloomPubDialog.tsx
@@ -11,7 +11,7 @@ import {
 } from "../../../react_components/BloomDialog/BloomDialog";
 import { DialogCancelButton } from "../../../react_components/BloomDialog/commonDialogComponents";
 import { useL10n } from "../../../react_components/l10nHooks";
-import { BloomCheckbox } from "../../../react_components/muiCheckBox";
+import { BloomCheckbox } from "../../../react_components/BloomCheckBox";
 import TextField from "@mui/material/TextField";
 import { WhatsThisBlock } from "../../../react_components/helpLink";
 import { BloomPalette } from "../../../react_components/color-picking/bloomPalette";
